+++ server/src/controllers/usersController.ts
@@ -1,14 +1,18 @@
 import { UserErrors } from "../errors";
-import { UserModel } from "../models/user";
+import { IUser, UserModel } from "../models/user";
 import bcrypt from "bcrypt";
-import { Express, Response } from "express";
+import jwt from "jsonwebtoken";
+import { Request, Response } from "express";
+import { TypedRequestBody } from "../types";
 
-export interface TypedRequestBody<T> extends Express.Request {
-  body: T
+type UserBody = {
+  username: string,
+  password: string,
+  isAdmin: boolean
 }
 
 async function registerUser(
-  req: TypedRequestBody<{ username: string, password: string, isAdmin: boolean }>,
+  req: TypedRequestBody<UserBody>,
   res: Response
 ): Promise<void> {
   const { username, password, isAdmin = false } = req.body;
@@ -31,4 +35,48 @@ async function registerUser(
   }
 }
 
-export default { registerUser }
+async function login(req: TypedRequestBody<Omit<UserBody, "isAdmin">>, res: Response): Promise<void> {
+  const { username, password } = req.body;
+
+  try {
+    const user: IUser = await UserModel.findOne({ username });
+
+    if (!user) {
+      res.status(400).json({ type: UserErrors.NO_USER_FOUND });
+      return
+    }
+
+    const isPasswordValid = await bcrypt.compare(password, user.password);
+    if (!isPasswordValid) {
+      res.status(400).json({ type: UserErrors.WRONG_CREDENTIALS });
+      return
+    }
+
+    const token = jwt.sign({ id: user._id, isAdmin: user.isAdmin }, "secret", {
+      expiresIn: '1 hour'
+    });
+
+    res.json({ token, userID: user._id });
+  } catch (err) {
+    res.status(500).json({ type: err });
+  }
+}
+
+async function availableMoney(req: Request, res: Response): Promise<void> {
+  const { userID } = req.params;
+
+  try {
+    const user = await UserModel.findById(userID);
+
+    if (!user) {
+      res.status(400).json({ type: UserErrors.NO_USER_FOUND });
+      return
+    }
+
+    res.json({ availableMoney: user.availableMoney });
+  } catch (err) {
+    res.status(500).json({ type: err });
+  }
+}
+
+export default { availableMoney, registerUser, login }
