+++ server/src/index.ts
@@ -3,7 +3,7 @@ import cors from 'cors'
 import mongoose from 'mongoose'
 require('dotenv').config();
 
-import { userRouter } from './routes/user';
+import { UserRouter } from './routes/user';
 import { ProductRouter } from './routes/product';
 
 const app = express();
@@ -11,7 +11,7 @@ const app = express();
 app.use(express.json());
 app.use(cors());
 
-app.use("/user", userRouter);
+app.use("/user", UserRouter);
 app.use("/products", ProductRouter);
 
 mongoose.connect(`mongodb+srv://pedrolportow:${process.env.DB_PASSWORD}@ecommerce-mern.4yui1uz.mongodb.net/ecommerce-mern`);
